Adds licenses; moves models.
